Greatly reduced warnings from PSScriptAnalyzer... maybe
Issues/Problems reduced in VS Code, but possibly not `PSScriptAnalyzer` as it isn't picking up the `PSScriptAnalyzerSettings.psd1` suppression file (inline suppressions didn't seem to help).

With the helpful side effect that it reduces the noise within the AI bundler output.

Removed loads of annoying whitespace. Probably need a script to do this as AI keeps screwing up and adding loads.